requirements.txt
flask
openai
faiss-cpu
numpy
